universe framework: extract vm params
