+++ universe/src/main/java/org/corfudb/universe/group/cluster/vm/RemoteOperationHelper.java
@@ -6,7 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.tools.ant.Project;
 import org.apache.tools.ant.taskdefs.optional.ssh.SSHExec;
 import org.apache.tools.ant.taskdefs.optional.ssh.Scp;
-import org.corfudb.universe.universe.vm.VmUniverseParams.Credentials;
+import org.corfudb.universe.api.deployment.vm.VmUniverseParams.Credentials;
 import org.corfudb.universe.util.IpAddress;
 
 import java.nio.file.Path;
