+++ universe/src/main/java/org/corfudb/universe/universe/vm/VmManager.java
@@ -18,6 +18,7 @@ import lombok.Getter;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.corfudb.common.result.Result;
+import org.corfudb.universe.api.deployment.vm.VmUniverseParams;
 import org.corfudb.universe.api.universe.UniverseException;
 import org.corfudb.universe.node.server.vm.VmCorfuServerParams.VmName;
 import org.corfudb.universe.universe.vm.ApplianceManager.ManagedEntityType;
