+++ universe/src/main/java/org/corfudb/universe/util/DockerManager.java
@@ -17,7 +17,7 @@ import lombok.Getter;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
-import org.corfudb.universe.api.deployment.DockerContainerParams;
+import org.corfudb.universe.api.deployment.docker.DockerContainerParams;
 import org.corfudb.universe.api.node.Node.NodeParams;
 import org.corfudb.universe.api.node.NodeException;
